Abstract out class serializers to their own files
